+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Features
 
 - Substitute `?` fields with `Option` fields in the converters related to NNS proposals.
+- Add retrieveBtcStatus to ckbtc minter canister.
 
 # Release.2023.11.21-1400Z